+++ lib/widgets/ar_view.dart
@@ -1,4 +1,5 @@
 import 'package:ar_flutter_plugin/managers/ar_anchor_manager.dart';
+import 'package:ar_flutter_plugin/managers/ar_location_manager.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:permission_handler/permission_handler.dart';
@@ -7,8 +8,11 @@ import 'package:ar_flutter_plugin/managers/ar_object_manager.dart';
 import 'package:ar_flutter_plugin/datatypes/config_planedetection.dart';
 
 // Type definitions to enforce a consistent use of the API
-typedef ARViewCreatedCallback = void Function(ARSessionManager arSessionManager,
-    ARObjectManager arObjectManager, ARAnchorManager arAnchorManager);
+typedef ARViewCreatedCallback = void Function(
+    ARSessionManager arSessionManager,
+    ARObjectManager arObjectManager,
+    ARAnchorManager arAnchorManager,
+    ARLocationManager arLocationManager);
 
 /// Factory method for creating a platform-dependent AR view
 abstract class PlatformARView {
@@ -43,7 +47,7 @@ createManagers(
     return;
   }
   arViewCreatedCallback(ARSessionManager(id, context, planeDetectionConfig),
-      ARObjectManager(id), ARAnchorManager(id));
+      ARObjectManager(id), ARAnchorManager(id), ARLocationManager());
 }
 
 /// Android-specific implementation of [PlatformARView]